C - Functions, nested loops

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	int l = 1, n = 2, s, count =2;
+	double l = 1, n = 2, s, count = 2;
 
 	printf ("%d, %d, ", l, n);
